+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - Update `ct tf reflow` to add `--single file [filename]` behaviour
 - Better diagnostic logging for `ct rest` and `ct az pim activate entra`
 - Add `ct az ad role definition|assignment browse|list` commands
+- Update `hicory-resolver` dependency
 
 # v0.35.1